% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_model.R
\name{group_model}
\alias{group_model}
\title{group_model}
\usage{
group_model(df)
}
\arguments{
\item{df}{dataframe with minimum columns 'Activity' and 'Time'}
}
\description{
This functions fits a linear model to 'Activity' as function of 'Time'
}
\examples{
group_model()
}
