% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_model.R
\name{hitachi_extract_lm}
\alias{hitachi_extract_lm}
\title{hitachi_extract_lm}
\usage{
hitachi_extract_lm(data)
}
\arguments{
\item{data}{tibble containing minimum columns 'Time', 'Concentration', 'Activity'}
}
\description{
This function fits linear models for each concentration and extracts model coefficients and statistics (glance, tidy, augment)
}
\examples{
hitachi_extract_lm()
}
