% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_model.R
\name{hitachi_extract_rate}
\alias{hitachi_extract_rate}
\title{hitachi_extract_rate}
\usage{
hitachi_extract_rate(data)
}
\arguments{
\item{data}{output from hitachi_extract_lm with one column tidy containing broom::tidy parameters}
}
\description{
This function extracts slopes and their std.errors from linear models per concentration
}
\examples{
hitachi_extract_rate()
}
