% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_read.R
\name{hitachi_read}
\alias{hitachi_read}
\title{hitachi_read}
\usage{
hitachi_read(file)
}
\arguments{
\item{file}{Hitachi output, interpreted by and saved as .txt}
}
\description{
This function allows you to import data from Hitachi LSC-8000 (.txt files) in tidy format.
}
\examples{
hitachi_read()
}
