% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_tibble.R
\name{hitachi_tibble}
\alias{hitachi_tibble}
\title{hitachi_tibble}
\usage{
hitachi_tibble(
  data,
  treatment,
  Join_bg = Join_bg,
  Join_sa = Join_sa,
  V_RuBP = 4e-06,
  C_RuBP = 6e-04,
  convert = TRUE
)
}
\arguments{
\item{data}{left_join(hitchi_read / hitachi_read_xls, design, by = Hitachi_ID) as input. Minimum columns 'Group' and 'CPM'.}

\item{treatment}{additional file providing groupwise information, e.g substrate concentration. Joined with data by column 'Group'.}

\item{Join_bg}{default to column name 'Join_bg', provides ties for background counts (e.g. dbl values).}

\item{Join_sa}{default to column name 'Join_sa', provides ties for specific activity (sa) counts (e.g. dbl values).}

\item{V_RuBP}{Volume [L] of RuBP-solution added to sa-reactions - default to 0.000004 .}

\item{C_RuBP}{Concentration [M] of RuBP-solution added to sa-reactions - default to 0.0006 .}

\item{convert}{default to T - converts CPM in Activity_nmol fixed carbon after groupwise background subtraction (Group == "control"). Requires parameters V_RuBP and C_RuBP!}
}
\description{
takes left_join(hitchi_read / hitachi_read_xls, design, by = Hitachi_ID) as
input and does background subtraction and sa-calculation on basis of provided
parameters V_RuBP and C_RuBP. If provided, adds additional groupwise information by
left_join(., treatment, by = 'Group').
}
\examples{
hitachi_tibble()
}
