% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_model.R
\name{model.nls1}
\alias{model.nls1}
\title{model.nls1}
\usage{
model.nls1(df)
}
\arguments{
\item{df}{}
}
\description{
This functions fits a Michaelis-Menten curve (non-linear least squares) to initial velocities
}
\examples{
model.nls1()
}
