% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_model.R
\name{model.nls2}
\alias{model.nls2}
\title{model.nls2}
\usage{
model.nls2(df)
}
\arguments{
\item{df}{}
}
\description{
This functions fits a simple inhibition model to initial velocities
}
\examples{
model.nls2()
}
